feat: add Open Graph tags
This commit adds Open Graph tags to the SEO component.

The Open Graph protocol enables any web page to become a rich object in a
social graph. For instance, this is used on Facebook to allow any web page to
have the same functionality as any other object on Facebook.

https://ogp.me/